general test code cleanup, removed duplicated code, improved tests structure
